Pull out heart room functions to a .h for easy diffing
